Fixes for SKILIKS-2272.

+++ protected/controllers/static/SimulationsController.php
@@ -40,12 +40,21 @@ class SimulationsController extends AjaxController implements AccountPageControl
             'status'      => Invite::STATUS_ACCEPTED
         ]);
 
-        // I remove this invites to roll back issue
-        foreach ($notUsedFullSimulations as $notUsedFullSimulation) {
-            $notUsedFullSimulation->delete();
+        // I remove more than 1 allowed to start lite sim {
+        if (1 < count($notUsedFullSimulations)) {
+            $i = 0;
+            foreach ($notUsedFullSimulations as $key => $notUsedFullSimulation) {
+                if (0 < $i) {
+                    $notUsedFullSimulation->delete();
+                    unset($notUsedFullSimulations[$key]);
+                }
+                $i++;
+            }
+            $notUsedFullSimulations = [];
         }
+        // I remove more than 1 allowed to start lite sim }
 
-        /*
+
         if (0 === count($notUsedFullSimulations)) {
             $newInviteForFullSimulation = new Invite();
             $newInviteForFullSimulation->owner_id = Yii::app()->user->data()->id;
@@ -54,13 +63,14 @@ class SimulationsController extends AjaxController implements AccountPageControl
             $newInviteForFullSimulation->lastname = Yii::app()->user->data()->profile->lastname;
             $newInviteForFullSimulation->scenario_id = $fullScenario->id;
             $newInviteForFullSimulation->status = Invite::STATUS_ACCEPTED;
+            $newInviteForFullSimulation->sent_time = time(); // @fix DB!
             $newInviteForFullSimulation->save(true, [
                 'owner_id', 'receiver_id', 'firstname', 'lastname', 'scenario_id', 'status'
             ]);
 
             $newInviteForFullSimulation->email = Yii::app()->user->data()->profile->email;
             $newInviteForFullSimulation->save(false);
-        }*/
+        }
         // check and add trial lite version }
 
         $this->render('simulations_personal', []);
@@ -81,12 +91,20 @@ class SimulationsController extends AjaxController implements AccountPageControl
             'status'      => Invite::STATUS_ACCEPTED
         ]);
 
-        // I remove this invites to roll back issue
-        foreach ($notUsedFullSimulations as $notUsedFullSimulation) {
-            $notUsedFullSimulation->delete();
+        // I remove more than 1 allowed to start lite sim {
+        if (1 < count($notUsedFullSimulations)) {
+            $i = 0;
+            foreach ($notUsedFullSimulations as $key => $notUsedFullSimulation) {
+                if (0 < $i) {
+                    $notUsedFullSimulation->delete();
+                    unset($notUsedFullSimulations[$key]);
+                }
+                $i++;
+            }
+            $notUsedFullSimulations = [];
         }
+        // I remove more than 1 allowed to start lite sim }
 
-        /*
         if (0 === count($notUsedFullSimulations)) {
             $newInviteForFullSimulation = new Invite();
             $newInviteForFullSimulation->owner_id = Yii::app()->user->data()->id;
@@ -95,46 +113,56 @@ class SimulationsController extends AjaxController implements AccountPageControl
             $newInviteForFullSimulation->lastname = Yii::app()->user->data()->profile->lastname;
             $newInviteForFullSimulation->scenario_id = $fullScenario->id;
             $newInviteForFullSimulation->status = Invite::STATUS_ACCEPTED;
+            $newInviteForFullSimulation->sent_time = time(); // @fix DB!
             $newInviteForFullSimulation->save(true, [
                 'owner_id', 'receiver_id', 'firstname', 'lastname', 'scenario_id', 'status'
             ]);
 
             $newInviteForFullSimulation->email = Yii::app()->user->data()->profile->email;
             $newInviteForFullSimulation->save(false);
-        }*/
+        }
         // check and add trial full version }
 
         // check and add trial lite version {
-        $fullScenario = Scenario::model()->findByAttributes(['slug' => Scenario::TYPE_LITE]);
+        $liteScenario = Scenario::model()->findByAttributes(['slug' => Scenario::TYPE_LITE]);
 
-        $notUsedFullSimulations = Invite::model()->findAllByAttributes([
+        $notUsedLiteSimulations = Invite::model()->findAllByAttributes([
             'receiver_id' => Yii::app()->user->data()->id,
-            'scenario_id' => $fullScenario->id,
+            'scenario_id' => $liteScenario->id,
             'email'       => Yii::app()->user->data()->profile->email,
             'status'      => Invite::STATUS_ACCEPTED
         ]);
 
-        // I remove this invites to roll back issue
-        foreach ($notUsedFullSimulations as $notUsedFullSimulation) {
-            $notUsedFullSimulation->delete();
+        // I remove more than 1 allowed to start lite sim {
+        if (1 < count($notUsedLiteSimulations)) {
+            $i = 0;
+            foreach ($notUsedLiteSimulations as $key => $notUsedLiteSimulation) {
+                if (0 < $i) {
+                    $notUsedLiteSimulation->delete();
+                    unset($notUsedLiteSimulations[$key]);
+                }
+                $i++;
+            }
+            $notUsedLiteSimulations = [];
         }
+        // I remove more than 1 allowed to start lite sim }
 
-        /*
-        if (0 === count($notUsedFullSimulations)) {
+        if (0 === count($notUsedLiteSimulations)) {
             $newInviteForFullSimulation = new Invite();
             $newInviteForFullSimulation->owner_id = Yii::app()->user->data()->id;
             $newInviteForFullSimulation->receiver_id = Yii::app()->user->data()->id;
             $newInviteForFullSimulation->firstname = Yii::app()->user->data()->profile->firstname;
             $newInviteForFullSimulation->lastname = Yii::app()->user->data()->profile->lastname;
-            $newInviteForFullSimulation->scenario_id = $fullScenario->id;
+            $newInviteForFullSimulation->scenario_id = $liteScenario->id;
             $newInviteForFullSimulation->status = Invite::STATUS_ACCEPTED;
+            $newInviteForFullSimulation->sent_time = time(); // @fix DB!
             $newInviteForFullSimulation->save(true, [
                 'owner_id', 'receiver_id', 'firstname', 'lastname', 'scenario_id', 'status'
             ]);
 
             $newInviteForFullSimulation->email = Yii::app()->user->data()->profile->email;
             $newInviteForFullSimulation->save(false);
-        }*/
+        }
         // check and add trial lite version }
 
         $this->render('simulations_corporate', []);
